% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amlr.driver}
\alias{amlr.driver}
\title{AMLR-preferred database driver}
\format{
A character vector of length one
}
\usage{
amlr.driver
}
\description{
A variable that contains the name of the preferred driver for connecting to
US AMLR SQL Server databases
}
\keyword{datasets}
