% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_wrappers.R
\name{amlr_wrappers}
\alias{amlr_wrappers}
\alias{amlr_box}
\alias{tableNA}
\title{Wrapper function}
\usage{
amlr_box(..., status = "warning", solidHeader = FALSE, collapsible = TRUE)

tableNA(...)
}
\arguments{
\item{...}{passed as-is to applicable base function}

\item{status}{passed to \code{\link[shinydashboard:box]{shinydashboard::box()}}. Default is \code{"warning"}}

\item{solidHeader}{passed to \code{\link[shinydashboard:box]{shinydashboard::box()}}. Default is \code{FALSE}}

\item{collapsible}{passed to \code{\link[shinydashboard:box]{shinydashboard::box()}}. Default is \code{TRUE}}
}
\description{
Wrapper functions with common AMLR defaults
}
\details{
\code{amlr_box} is a wrapper around \code{\link[shinydashboard:box]{shinydashboard::box()}},
with the following defaults: \code{status = "warning"}, \code{solidHeader = FALSE},
\code{collapsible = TRUE}. The user can override these individually, or simply
use \code{\link[shinydashboard:box]{shinydashboard::box()}} itself.

\code{tableNA} is a wrapper around the \code{\link[base:table]{base::table()}} function,
with the \code{useNA} argument set to 'ifany'\verb{. Specifically, run the }table(..., useNA = 'ifany')`.
}
\examples{
tableNA(c(1, 2, NA, 2, NA, 2, 1, 3))
tableNA(c(1, 2, 2, 2, 1, 3))

}
