% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrian-package.R
\docType{package}
\name{amlrian-package}
\alias{amlrian-package}
\alias{amlrian}
\title{Functions for NOAA US AMLR Program}
\description{
Functions for NOAA US AMLR Program
}
\details{
Functions for NOAA US AMLR Program
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{internal}
