% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database.R
\name{mod_database}
\alias{mod_database}
\alias{mod_database_ui}
\alias{mod_database_server}
\title{Database connection module for AMLR shiny apps}
\usage{
mod_database_ui(
  id,
  col.width = 5,
  server.default = NULL,
  database.default = NULL,
  port.default = NULL,
  conn.default = "trusted",
  uid.default = NULL,
  pwd.default = NULL
)

mod_database_server(id, pool.list = list(), filedsn = NULL)
}
\arguments{
\item{id}{module namespace, see \code{\link[shiny:NS]{shiny::NS()}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{server.default}{default character value for "Server Hostname" widget;
default is \code{NULL}. If \code{NULL}, the module defaults to
\code{paste0(Sys.info()[["nodename"]], "\\\\SQLEXPRESS")}}

\item{database.default}{default character value for "Database" widget;
default is \code{NULL}}

\item{port.default}{default numeric value for "Port number" widget; default
is \code{NULL}}

\item{conn.default}{default character value for "Connection type" widget;
default is \code{"trusted"}}

\item{uid.default}{default character value for "User" widget; default is
\code{NULL}}

\item{pwd.default}{default character value for "Password" widget; default is
\code{NULL}}

\item{pool.list}{a (named) list of pool objects created by \code{\link[pool:dbPool]{pool::dbPool()}}.
The names of the objects in this list will appear as
'database connection' \code{\link[shiny:radioButtons]{shiny::radioButtons()}} options.}

\item{filedsn}{character; default is \code{NULL.} The file path to a DSN file with a
database connection. If not \code{NULL}, Tamatoa will try to try to establish a
database connection using \link[pool:dbPool]{pool::dbPool}. See 'Details' for an example}
}
\value{
Returns a reactive of the pool connection specified by the user, which can
then be passed to other modules.
}
\description{
Database connection module for AMLR shiny apps
}
\details{
This module includes the UI for a user to specify and
create a database connection. It also allows users to connect to the database
of their choice in a parent function, and pass that connection (those
connections) via \code{pool.list}. Additionally, users can provide a filedsn,
and the module will attempt to create the connection defined by the \code{filedsn}
argument.

If \code{filedsn} argument is not \code{NULL}, then Tamatoa will try to make a
connection via:

\code{pool::dbPool(odbc::odbc(), filedsn = filedsn)}

The order of priority for the default selected database: \code{pool.list},
\code{filedsn}, then 'Other'.
}
