% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_test.R
\name{shiny_test}
\alias{shiny_test}
\title{Open the package test Shiny app}
\usage{
shiny_test(..., filedsn = NULL)
}
\arguments{
\item{...}{arguments passed as a list to the \code{options} argument of
\code{\link[shiny:runApp]{shiny::runApp()}}}

\item{filedsn}{character; see \code{\link[=mod_database]{mod_database()}} for details}
}
\description{
Open the package test Shiny app
}
\details{
The Shiny app is a testing/development app. To date, is has
been used to develop and test the database connection and output modules
contained in this package
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
shiny_test(launch.browser = TRUE)
\dontshow{\}) # examplesIf}
}
